feat: Separate the container init to an abstract helper class
