+++ ors-test-scenarios/src/test/java/integrationtests/EnvironmentChecks.java
@@ -4,11 +4,7 @@ import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
-import org.testcontainers.containers.wait.strategy.WaitStrategy;
-import org.testcontainers.images.builder.ImageFromDockerfile;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.testcontainers.junit.jupiter.TestcontainersExtension;
 import org.testcontainers.shaded.com.fasterxml.jackson.databind.JsonNode;
@@ -16,10 +12,7 @@ import utils.OrsApiRequests;
 import utils.OrsContainerFileSystemCheck;
 
 import java.io.IOException;
-import java.nio.file.Path;
-import java.time.Duration;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Stream;
 
 import static utils.OrsApiRequests.checkAvoidAreaRequest;
@@ -28,75 +21,20 @@ import static utils.OrsApiRequests.checkAvoidAreaRequest;
 @ExtendWith(TestcontainersExtension.class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class EnvironmentChecks {
-    private final Map<String, String> defaultEnv = Map.of(
-            "logging.level.org.heigit", "INFO",
-            "ors.engine.graphs_data_access", "MMAP"
-    );
-
-    private GenericContainer<?> warContainer;
-    private GenericContainer<?> jarContainer;
+public class EnvironmentChecks extends ContainerInitializer {
 
     static Stream<Object[]> data() {
         return Stream.of(
-                new Object[]{"ors-test-scenarios-war", "3.12"},
-                new Object[]{"ors-test-scenarios-jar", "3.13"}
+                new Object[]{ContainerTestImage.WAR_CONTAINER},
+                new Object[]{ContainerTestImage.JAR_CONTAINER}
         );
     }
 
-    private GenericContainer<?> initContainer(String baseImage) {
-        WaitStrategy waitStrategy = new HttpWaitStrategy()
-                .forPort(8080)
-                .forStatusCode(200)
-                .forPath("/ors/v2/health")
-                .withStartupTimeout(Duration.ofSeconds(80));
-
-        GenericContainer<?> container = new GenericContainer<>(
-                new ImageFromDockerfile(baseImage, false)
-                        .withFileFromPath("ors-api", Path.of("../ors-api"))
-                        .withFileFromPath("ors-engine", Path.of("../ors-engine"))
-                        .withFileFromPath("ors-report-aggregation", Path.of("../ors-report-aggregation"))
-                        .withFileFromPath("pom.xml", Path.of("../pom.xml"))
-                        .withFileFromPath("ors-config.yml", Path.of("../ors-config.yml"))
-                        .withFileFromPath("Dockerfile", Path.of("../ors-test-scenarios/src/test/resources/Dockerfile"))
-                        .withFileFromPath(".dockerignore", Path.of("../.dockerignore"))
-                        .withTarget(baseImage)
-        )
-                .withEnv(defaultEnv)
-                .withFileSystemBind("./graphs-integrationtests", "/home/ors/openrouteservice/graphs", BindMode.READ_WRITE)
-                .withExposedPorts(8080)
-                .withLogConsumer(outputFrame -> System.out.print(outputFrame.getUtf8String()))
-                .waitingFor(waitStrategy);
-
-        if (baseImage.equals("ors-test-scenarios-war")) {
-            if (warContainer == null || !warContainer.isRunning()) {
-                warContainer = container;
-                warContainer.start();
-            }
-            return warContainer;
-        } else {
-            if (jarContainer == null || !jarContainer.isRunning()) {
-                jarContainer = container;
-                jarContainer.start();
-            }
-            return jarContainer;
-        }
-    }
-
-    @AfterEach
-    public void resetEnv() {
-        if (warContainer != null) {
-            warContainer.withEnv(defaultEnv);
-        }
-        if (jarContainer != null) {
-            jarContainer.withEnv(defaultEnv);
-        }
-    }
 
     @Order(1)
     @MethodSource("data")
     @ParameterizedTest(name = "{0}")
-    void testBuildAllImagesAndGraphs(String targetImage) throws IOException, InterruptedException {
+    void testBuildAllImagesAndGraphs(ContainerTestImage targetImage) throws IOException, InterruptedException {
         GenericContainer<?> container = initContainer(targetImage);
         container.addEnv("ors.engine.profiles.public-transport.enabled", "false");
         container.addEnv("ors.engine.profile_default.enabled", "true");
@@ -139,11 +77,11 @@ public class EnvironmentChecks {
     @Order(2)
     @MethodSource("data")
     @ParameterizedTest(name = "{0}")
-    void testAvoidAreaRequestAndGeoToolsPopulation(String targetImage) throws IOException, InterruptedException {
+    void testAvoidAreaRequestAndGeoToolsPopulation(ContainerTestImage targetImage) throws IOException, InterruptedException {
         GenericContainer<?> container = initContainer(targetImage);
 
         String geoToolsPath;
-        if (container.getDockerImageName().contains("ors-test-scenarios-war"))
+        if (targetImage.equals(ContainerTestImage.WAR_CONTAINER))
             geoToolsPath = "/usr/local/tomcat/temp/GeoTools";
         else geoToolsPath = "/tmp/GeoTools";
 
@@ -155,7 +93,7 @@ public class EnvironmentChecks {
     @Order(3)
     @MethodSource("data")
     @ParameterizedTest(name = "{0}")
-    void testTwoProfilesActivatedByEnv(String targetImage) throws IOException {
+    void testTwoProfilesActivatedByEnv(ContainerTestImage targetImage) throws IOException {
         GenericContainer<?> container = initContainer(targetImage);
 
         // Activate two new profiles alongside the default driving-car profile
